テーマモード変更ボタンの追加とOSSライセンス表示ボタンを単独ファイルへ移動
